Rename version3 to version4 (sync with hpctoolkit version)
