+++ edu.rice.cs.hpcdata/src/edu/rice/cs/hpcdata/db/version4/DataRecord.java
@@ -1,4 +1,4 @@
-package edu.rice.cs.hpcdata.db.version3;
+package edu.rice.cs.hpcdata.db.version4;
 
 /***
  * struct trace record
