+++ edu.rice.cs.hpcdata/src/edu/rice/cs/hpcdata/experiment/scope/RootScope.java
@@ -19,8 +19,8 @@ import java.io.File;
 import java.io.IOException;
 
 import edu.rice.cs.hpcdata.db.MetricValueCollectionWithStorage;
-import edu.rice.cs.hpcdata.db.version3.DataSummary;
-import edu.rice.cs.hpcdata.db.version3.MetricValueCollection3;
+import edu.rice.cs.hpcdata.db.version4.DataSummary;
+import edu.rice.cs.hpcdata.db.version4.MetricValueCollection3;
 import edu.rice.cs.hpcdata.experiment.BaseExperiment;
 import edu.rice.cs.hpcdata.experiment.extdata.IThreadDataCollection;
 import edu.rice.cs.hpcdata.experiment.metric.IMetricValueCollection;
