+++ edu.rice.cs.hpcremote/src/edu/rice/cs/hpcremote/data/DecompressionThread.java
@@ -10,7 +10,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.zip.InflaterInputStream;
 
 import edu.rice.cs.hpctraceviewer.data.ImageTraceAttributes;
-import edu.rice.cs.hpcdata.db.version3.DataRecord;
+import edu.rice.cs.hpcdata.db.version4.DataRecord;
 import edu.rice.cs.hpcdata.util.CallPath;
 import edu.rice.cs.hpctraceviewer.data.timeline.ProcessTimeline;
 import edu.rice.cs.hpctraceviewer.data.util.Constants;
