+++ edu.rice.cs.hpctraceviewer.data/META-INF/MANIFEST.MF
@@ -18,4 +18,4 @@ Export-Package: edu.rice.cs.hpctraceviewer.data,
  edu.rice.cs.hpctraceviewer.data.timeline,
  edu.rice.cs.hpctraceviewer.data.util,
  edu.rice.cs.hpctraceviewer.data.version2,
- edu.rice.cs.hpctraceviewer.data.version3
+ edu.rice.cs.hpctraceviewer.data.version4
