+++ edu.rice.cs.hpctraceviewer.data/src/edu/rice/cs/hpctraceviewer/data/TraceDataByRank.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.Vector;
 
-import edu.rice.cs.hpcdata.db.version3.DataRecord;
+import edu.rice.cs.hpcdata.db.version4.DataRecord;
 import edu.rice.cs.hpcdata.util.Constants;
 import edu.rice.cs.hpctraceviewer.data.version2.AbstractBaseData;
