+++ edu.rice.cs.hpctraceviewer.data/src/edu/rice/cs/hpctraceviewer/data/local/LocalDBOpener.java
@@ -15,7 +15,7 @@ import edu.rice.cs.hpcdata.util.Util;
 import edu.rice.cs.hpctraceviewer.data.AbstractDBOpener;
 import edu.rice.cs.hpctraceviewer.data.DatabaseAccessInfo;
 import edu.rice.cs.hpctraceviewer.data.SpaceTimeDataController;
-import edu.rice.cs.hpctraceviewer.data.version3.FileDB3;
+import edu.rice.cs.hpctraceviewer.data.version4.FileDB4;
 
 
 /****************************************************
@@ -94,7 +94,7 @@ public class LocalDBOpener extends AbstractDBOpener
 			break;
 		case 3:
 		case Constants.EXPERIMENT_SPARSE_VERSION:
-			fileDB = new FileDB3();
+			fileDB = new FileDB4();
 			break;
 		default:
 			throw new InvalExperimentException("Trace data version is not unknown: " + version);
